Enumerable::ENUMERABLE_SIZE
Associated constant for the number of values in this enumerable.
If the number exceeds `usize::MAX`, accessing the constant fails
at compile time.

This is useful to create a compact storage to map all the values:

```
let mut data: [None; MyEnum::ENUMERABLE_SIZE];
```

I chose name `ENUMERABLE_SIZE` not `SIZE` to avoid confusion with
possible other meaning of `SIZE` (for example, `for `u32` `SIZE`
could also mean size in bytes).

Follow-up to this PR would be implementing:
- `Enumerable::enumerable_index` to map enumerable value to `usize` index
- `Enumerable::enumerable_from_index` for the opposite operation