add new methods for trait `Enumerable`
- `enumerator_since`: create an enumerator since a given index.
- `enumerable_from_index`: get element from index.

`enumerable_index_of` is not added because it requires `PartialEq`. Adding a method with type bound and default implementation is possible, but it's nearly impossible for derive macro to handle it correctly (as it can not know whether the implementing type also implements `PartialEq`). Maybe a separate trait is a good solution but I'm now quite sure if it's worthwhile.